Improve vlog GC logic
The current implementation would sample a vlog even when we knew how much
data can be discarded from it. The sampling might not find enough data
to discard even when it is there.  The proposed implementation directly
rewrites the value log file if we've picked it from discard stats map.

In case if the discard stats map doesn't have any file then we pick one
randomly.